refactor: remove compiled library. Now header only
