+++ src/cpp/pss/astrotypes/sigproc/CMakeLists.txt
@@ -1,7 +1,4 @@
 set(module_sigproc_lib_src_cpu
-    src/Header.cpp
-    src/HeaderField.cpp
-    src/SigProcVariable.cpp
     PARENT_SCOPE
 )
