+++ src/cpp/pss/astrotypes/sigproc/detail/SigProcVariable.cpp
@@ -56,6 +56,61 @@ unsigned SigProcVariable<boost::units::quantity<Unit, NumericalRep>>::read(std::
     return stream.gcount();
 }
 
+// ------- std::string ------------------------
+inline unsigned SigProcVariable<std::string>::write(std::ostream& stream, std::string const& var)
+{
+    int32_t size=static_cast<int32_t>(var.size());
+    if( size > 80 ) {
+        throw std::runtime_error("SigProc::write: illegal size of string (max 80): " + std::to_string(size));
+    }
+    // write out the size of the string
+    unsigned s = SigProcVariable<decltype(size)>::write(stream, size);
+    stream.write(var.data(), size);
+    return size + s;
+}
+
+inline unsigned SigProcVariable<std::string>::read(std::istream& stream, std::string& var)
+{
+    // read in the size of the string
+    int32_t size=0;
+    stream.read(reinterpret_cast<char*>(&size), sizeof(size));
+    auto s = sizeof(int32_t);
+
+    if( size < 0 || size > 80 ) {
+        std::runtime_error e("SigProcVariable<std::string>::read: illegal size of string: " + std::to_string(size));
+        throw e;
+    }
+
+    if(!stream.fail()) {
+        var.resize(size);
+        stream.read(&var[0], size);
+        s += stream.gcount();
+    }
+    return s;
+}
+
+inline void SigProcLabel::write(std::ostream& stream) const
+{
+    SigProcVariable<std::string>::write(stream, _string);
+}
+
+inline void SigProcLabel::read(std::istream& stream)
+{
+    SigProcVariable<std::string>::read(stream, _string);
+}
+
+inline std::istream& operator>>(std::istream& stream, SigProcLabel& var)
+{
+    var.read(stream);
+    return stream;
+}
+
+inline std::ostream& operator<<(std::ostream& stream, SigProcLabel const& var)
+{
+    var.write(stream);
+    return stream;
+}
+
 } // namespace sigproc
 } // namespace astrotypes
 } // namespace pss
