+++ src/cpp/pss/astrotypes/sigproc/examples/CMakeLists.txt
@@ -4,4 +4,4 @@ set(examples_src
 
 add_library("sigproc_examples" ${examples_src})
 add_executable("sigproc_cat" src/sigproc_cat.cpp)
-target_link_Libraries(sigproc_cat ${ASTROTYPES_LIBRARIES})
+target_link_Libraries(sigproc_cat)
