+++ src/cpp/pss/astrotypes/sigproc/test/CMakeLists.txt
@@ -9,5 +9,5 @@ set(
 
 add_executable(gtest_sigproc ${gtest_sigproc_src} )
 #target_link_libraries(gtest_sigproc ${ASTROTYPES_TEST_UTILS} ${ASTROTYPES_LIBRARIES} ${GTEST_LIBRARIES})
-target_link_libraries(gtest_sigproc ${ASTROTYPES_TEST_UTILS} ${ASTROTYPES_LIBRARIES} ${GTEST_LIBRARIES})
+target_link_libraries(gtest_sigproc ${ASTROTYPES_TEST_UTILS} ${GTEST_LIBRARIES})
 add_test(gtest_sigproc gtest_sigproc)
